+++ .github/workflows/go.yml
@@ -44,12 +44,9 @@ jobs:
 
       - name: Setup Extra
         run: |
-          go get -u golang.org/x/lint/golint
           go get -u github.com/securego/gosec/v2/cmd/gosec@v2.8.0
           go get ./...
 
-      - name: Lint
-        run: make lint
 
 #      Ignoring until I write tests
 #      - name: Test
