+++ .github/workflows/golanglint-ci.yml
@@ -1,0 +1,38 @@
+name: Golanglint CI
+
+on:
+  push:
+    branches:
+      - '**'
+    tags-ignore:
+      - '**'
+    paths:
+      - 'go.sum'
+      - 'go.mod'
+      - '**.go'
+      - .github/workflows/golangci-lint.yml
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  Go:
+    runs-on: ubuntu-latest
+    env:
+      GO111MODULE: on
+
+    steps:
+      - name: Install Go
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.17
+
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: golangci-lint
+        uses: golangci/golangci-lint-action@v2
+        with:
+          version: latest
+          args: -E revive,gofmt
+          skip-go-installation: true
