+++ cmd/.goreleaser.yml
