+++ go.mod
@@ -1,14 +1,37 @@
 module github.com/Cyb3r-Jak3/notification_relay
 
-go 1.16
+go 1.17
 
 require (
-	github.com/Cyb3r-Jak3/common v0.1.4
 	github.com/JohannesKaufmann/html-to-markdown v1.3.0
-	github.com/google/go-github/v35 v35.3.0
 	github.com/mcuadros/go-defaults v1.2.0
 	github.com/nickname32/discordhook v1.0.2
 	github.com/sirupsen/logrus v1.8.1
 	github.com/urfave/cli/v2 v2.3.0
 	golang.org/x/oauth2 v0.0.0-20210514164344-f6687ab2804c
+)
+
+require (
+	github.com/PuerkitoBio/goquery v1.5.1 // indirect
+	github.com/andersfylling/snowflake v1.3.0 // indirect
+	github.com/andybalholm/cascadia v1.1.0 // indirect
+	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
+	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/json-iterator/go v1.1.10 // indirect
+	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
+	github.com/modern-go/reflect2 v0.0.0-20180701023420-4b7aa43c6742 // indirect
+	github.com/russross/blackfriday/v2 v2.0.1 // indirect
+	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110 // indirect
+	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.25.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
+
+require (
+	github.com/Cyb3r-Jak3/common/v3 v3.0.0-20211021163845-fe6144e9adfb
+	github.com/google/go-github/v39 v39.2.0
+	github.com/google/go-querystring v1.1.0 // indirect
+	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
 )
